requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cssselect==1.3.0
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.30.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.2
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.1
sympy==1.13.3
thinc==8.2.5
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.2.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2